Remove the settings struct and its trait
The concept of `allowed_user` and `allowed_groups` can't be applied when
inspecting child objects.

For example, when a Deployment is created, the resulting Pod creation
request is not seen as done by the user who created the Deployment
object, but rather it's seen as created by some internal k8s service
account.