Return a Vec when a resource is requested by name
Depending on the namespace filter, it could happen that ingresses and
services could be duplicated with the same name on different
namespaces. Provide a signature for this methods that resemble this
option in a natural way.

The Vec could be empty, in which case no resource was found with that
name, one element or more than one element if a resource with the same
name exists on different namespaces and the consumer call is not
filtering on a specific namespace.