Fix k8s-openapi dependency to break building of rust policies
Starting from k8s-openapi v0.14, it is NOT recommended to be explicit about
the kubernetes features to be used when building a library. That's because
the final version of the k8s API to be supported must be made by the consumer
of the library.

Otherwise it's possible to end up in situations like:
* kubewarden-policy-sdk enables the `v1_22` feature of k8s-openapi
* `consumer-foo` requires k8s-openapi too but enables the `v1_23` feature
A build error is then raised by cargo because the same version must be picked
by all the crates making use of k8s-openapi.

Because of that, no feature is chosen inside of the `dependencies` section.
This however can lead to issues when executing commands like
cargo `build|check|doc`. That's because the `k8s-openapi` is specified again
inside of the `dev-dependencies`, this time with a k8s feature enabled

Signed-off-by: Flavio Castelli <fcastelli@suse.com>