fix: handle default features being disabled
When the default features are disabled, the k8s-openapi crate is no
longer required.

The recent helpers introduced to deal with Kubernetes resources
containing Pod didn't take that into account.

That causes the library to not compile properly when the imported
without the default features.

This commit fixes this and extends our unit tests to ensure this won't
happen again.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>