fix: CertificateVerificationRequest{}.not_after as None if empty
CertificateVerificationRequest{}.not_after is an Option. if Some,
it is taken into account to check the certificate against the date,
here:
https://github.com/kubewarden/policy-evaluator/blob/71270655862d9f53bc23feafb3c4e9b4ed982bef/src/callback_handler/crypto.rs#L52-L54

Still, since this is a callback, it may happen that the guest caller
provides a `CertificateVerificationRequest{}` where its
`CertificateVerificationRequest{}.not_after` is present, yet set to the
empty string.
It would then deserialize to `Some("")` instead of `None`. This would
incorrectly skip the expiration check.

Hence, ensure that on (de)serialization on the host, we consider empty
strings as `None`.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>