refactor!: sigstore verifier
Change the Sigstore verifiers to use `BTreeMap` instead of `HashMap` to
store annotations.

This is required by a change done by the latest version of the
sigstore-rs crate.

This refactor breaks our Rust API, but it doesn't change the waPC one.
These are just two different data structures used to represent hashes

Signed-off-by: Flavio Castelli <fcastelli@suse.com>